Sprint 5 - 12.21.2017 - MU
Changed filters picker views to be buttons that popup view controller

began loading switches to be on/off based on Db value for muteMode and mealPlan. Also, began the code for sending data to the database depending on if switch is on/off when edit profile is saved.

Did not get to change profile picture feature like I had hoped. Will do that tomorrow.